Move StepRecorder implementation to a separate module for better organization
